+++ CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-None
+### Added
+
+- Added a configuration option `ignoreWarningMessages` in order to be able to ignore the following warning messages (especially if they come repeatedly, see #4):
+  - "Found multiple files ... will use ..."
+  - "... file does not exist. Ignoring ..."
+
+### Fixed
+
+- Improve warning message if the credo output is empty
+- Try to capture only the JSON-part of the credo output if it includes info messages from mix, for instance.
 
 ## [0.1.1] - 2020-11-24
 
@@ -21,7 +30,7 @@ None
 
 - Use credo issues' `trigger`s to mark the correct substring (#2)
 
-## Changed
+### Changed
 
 - Display `'refactor'` issues as `Information` instead of hints (#3).
   This allows the issues to be displayed more clearly and also shown in the `Problems` panel.